haystack/requirements.txt
streamlit
python-dotenv
sentence-transformers
chromadb
google-generativeai
PyPDF2
tqdm
langchain
requests
pdfplumber
